Move startup-log crash-tail attach out of shared scrubber (cmux-cli target has no StartupBreadcrumbLog)
